[#90] refactor: ColumnDelegate 클래스, 객체를 삭제하고 ColumnViewController가 		UITableViewDelegate의 역할을 하도록 채택하고 이전의 코드를 다 적용함
	이유: 기존 Delegate객체는 이벤트가 발생한 경우 노티를 보내는 역할만 했었고,
		특히 노티에 대한 옵저버가 ColumnViewController 하나밖에 없었다.
		원래 노티피케이션이라는게 1:n관계에서 여러 옵저버에게 보낼때 적합한데
		옵저버가 하나라서 아쉽고(비효율적이다), Delegate 객체도 다른 역할(모델객체를 생성한다거나)은
		하지 않아서 아쉽다(비효율적이다).
		따라서 기존의 델리게이트 객체는 삭제하고, ColumnViewController 객체가 Delegate의 역할을 하여
		좀 더 간단하고 유연한 구조가 되게 만들었다.
